pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-finance-env"
version = "1.0.0"
description = "Personal Finance OpenEnv API"
requires-python = ">=3.8"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi==0.110.0",
    "uvicorn==0.29.0",
    "pydantic==2.6.4",
    "openai==1.14.2",
    "httpx==0.27.0",
    "jsonref==1.1.0",
    "requests==2.31.0",
    "pyyaml==6.0.1",
    "jinja2==3.1.3",
    "python-multipart==0.0.9"
]

[project.scripts]
server = "server.app:main"
